+++ demo/src/main/java/com/example/demo/service/CompraService.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface CompraService {
     void crearListaCompras(List<ComprasEntity> listaCompras);
 
-    void deleteCompras(Integer idCompras);
+    void deleteCompras(Long idCompras);
 }
